requirements.txt
asyncpg
aiogram
python-decouple
APScheduler
psycopg2-binary